+++ tamr_client/attribute/_attribute.py
@@ -50,7 +50,7 @@ class ReservedName(TamrClientException):
     pass
 
 
-def from_resource_id(session: Session, dataset: Dataset, id: str) -> Attribute:
+def by_resource_id(session: Session, dataset: Dataset, id: str) -> Attribute:
     """Get attribute by resource ID
 
     Fetches attribute from Tamr server